pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-wow"
version = "1.0.0"
description = "AI Way of Working - Standardized AI-assisted software engineering framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "AI Way of Working Contributors" }
]
keywords = [
    "ai",
    "copilot",
    "claude",
    "developer-experience",
    "coding-standards",
    "framework"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
ai-wow = "ai_wow.cli:main"

[project.urls]
Homepage = "https://github.com/soydachi/ai-wow"
Documentation = "https://github.com/soydachi/ai-wow#readme"
Repository = "https://github.com/soydachi/ai-wow"
Issues = "https://github.com/soydachi/ai-wow/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_wow"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ai_wow"
